fix(prettier-plugin-jsdoc): format type with Prettier on format step
Originally, the type formatting with Prettier was on the prepare step and it was always
executed; after this commit, there's an option controlling it and it happens on the "format
types" step. The reason for the change was that the formatting of the dot for objects and
arrays and the changes for string literals couldn't happen before Prettier, otherwise those
possible changes would be discarded.