fix(prettier-plugin-jsdoc): move the example tag order to the end
BREAKING CHANGE: The example tag order has been changed to a lower position.
This change was made in order to fix a bug in which TypeScript would detect
'any' if the tag was before param, property and template. This is breaking
because saving a file with this change would cause possible unrelated blocks
to change.