fix: move return tag to tags with desc as name
Before moving the return tag to the new group, even on inline mode, it was
rendering the description in a column format.
Now, it will correctly move it to a new line in inline mode, and in column
mode, it will align it with the name property, instead of the description,
so the white space will be reduced.